+++ CHANGES.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## Unreleased
+
+### Modified
+
+- hol2dk theorem now splits a term_abbrevs file into several files
+  if it is too big
+
 ## 1.0.0 (2024-02-25)
 
 ### Added