+++ CHANGES.md
@@ -7,8 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Modified
 
-- hol2dk theorem now splits a term_abbrevs file into several files
-  if it is too big
+- hol2dk theorem now splits a term_abbrevs file into several files if
+  it is too big. The option --max-abbrevs allows to change the maximum
+  number of definitions in a term abbreviations file.
 
 ## 1.0.0 (2024-02-25)