+++ CHANGES.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
-## Unreleased
+## 2.0.0 (2024-04-23)
 
 Big improvements in Lambdapi and Coq file generation time, and Coq checking time.